+++ source/MP4.ELST.cpp
@@ -45,8 +45,8 @@ std::string ELST::getContent( void )
     o << "Entry Count: " << this->_entryCount << "\n";
     o << "Entries:\n";
     o << "Segment Duration   Time   Rate\n";
-    int count = _entries.size();
-    int omitted = 0;
+    long count = _entries.size();
+    long omitted = 0;
     if( count > 20 )
     {
         omitted = _entries.size() - 20;
